refactor: re-organize routes
