+++ main.ts
@@ -1,8 +1,5 @@
 import { serve } from "https://deno.land/std@0.153.0/http/mod.ts";
-import { blog } from "./blog.ts";
-import { news } from "./news.ts";
-import { release } from "./releases.ts";
-import { status } from "./status.ts";
+import { blog, news, release, status } from "./routes.ts";
 
 serve(async (req) => {
   const path = new URL(req.url).pathname.substring(1);
